+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.28.1](https://github.com/fabiocaccamo/python-fontbro/releases/tag/0.28.1) - 2026-09-13
 -   Update `set_name` to not add legacy Mac name records to fonts that don't have them (Mac records are updated only if already present).
 -   Fix `get_names` returning names in other languages (eg. japanese) instead of the english ones returned by `get_name`.
 -   Fix `get_name` returning `None` when the font has no Windows English name records (fallback to the other name records).